Don't try to delete docs directory if it does not exist.

+++ build.cake
@@ -58,7 +58,8 @@ Task("UpdateDocs")
 	.Does(() =>
 	{
 		docsDirectory = new DirectoryPath(docsBranchName);
-		DeleteDirectory(docsDirectory, new DeleteDirectorySettings { Recursive = true, Force = true });
+		if (DirectoryExists(docsDirectory))
+			DeleteDirectory(docsDirectory, new DeleteDirectorySettings { Recursive = true, Force = true });
 		GitClone(docsRepoUri, docsDirectory, new GitCloneSettings { BranchName = docsBranchName });
 
 		Information($"Updating documentation at {docsDirectory}.");
